third_party/blink/web_tests/fast/peerconnection/RTCPeerConnection-generateCertificate-expected.txt

Tests RTCPeerConnection.generateCertificate RSA/ECDSA.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS generateCertificate({ name: "RSASSA-PKCS1-v1_5", modulusLength: 2048, publicExponent: new Uint8Array([1, 0, 1]), hash: "SHA-256" })
PASS certRSA is non-null.
PASS new Date().getTime() < certRSA.expires is true
PASS certRSA.getFingerprints().length == 1 is true
PASS certRSA.getFingerprints()[0].algorithm === "sha-256" is true
PASS isFingerprintValue(certRSA.getFingerprints()[0].value) is true
PASS certRSA.getFingerprints()[0] != certRSA.getFingerprints()[0] is true
PASS generateCertificate({ name: "ECDSA", namedCurve: "P-256" })
PASS certECDSA is non-null.
PASS new Date().getTime() < certECDSA.expires is true
PASS certECDSA.getFingerprints().length == 1 is true
PASS certECDSA.getFingerprints()[0].algorithm === "sha-256" is true
PASS isFingerprintValue(certECDSA.getFingerprints()[0].value) is true
PASS certECDSA.getFingerprints()[0] != certECDSA.getFingerprints()[0] is true
PASS certExpiresNegativeOneDay is non-null.
PASS new Date().getTime() < certExpiresNegativeOneDay.expires is true
PASS certExpiresNegativeOneDay.getFingerprints().length == 1 is true
PASS certExpiresNegativeOneDay.getFingerprints()[0].algorithm === "sha-256" is true
PASS isFingerprintValue(certExpiresNegativeOneDay.getFingerprints()[0].value) is true
PASS certExpiresNegativeOneDay.getFingerprints()[0] != certExpiresNegativeOneDay.getFingerprints()[0] is true
PASS certExpiresZero is non-null.
PASS new Date().getTime() >= certExpiresZero.expires is true
PASS certExpiresPositiveOneDay is non-null.
PASS Math.abs(certExpiresPositiveOneDay.expires - (new Date().getTime() + dayInMs)) <= minuteInMs is true
PASS successfullyParsed is true

TEST COMPLETE
